pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "mail-mcp"
version = "0.1.0"
description = "Generic IMAP+SMTP MCP server — intent-first email access for AI agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "KpihX", email = "ivann.kamdem-pouokam@polytechnique.edu" }]
keywords = ["mcp", "email", "imap", "smtp", "ai", "agent"]

dependencies = [
    "fastmcp>=2.0",
    "imapclient>=3.0",
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "html2text>=2024.2",
    "python-dotenv>=1.0",
]

[project.scripts]
mail-mcp = "mail_mcp.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
